+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 6.1.0 (in progress)
 
 ##### New Features
-* [#1851](https://github.com/oshi/oshi/pull/1851): Add PhysicalProcessor class to expose hybrid processor topology - [@dbwiddis](https://github.com/dbwiddis).
+* [#1851](https://github.com/oshi/oshi/pull/1851),
+  [#1858](https://github.com/oshi/oshi/pull/1858): Add PhysicalProcessor class to expose hybrid processor topology - [@dbwiddis](https://github.com/dbwiddis).
 
 ##### Bug fixes / Improvements
 * [#1831](https://github.com/oshi/oshi/pull/1831): Improve Solaris and AIX process listing using procfs - [@dbwiddis](https://github.com/dbwiddis).